Added admin and generate report views
