+++ cs_app/views/__init__.py
@@ -1,2 +1,3 @@
 from .admin_views import *
-from . authorization_views import *
+from .authorization_views import *
+from .generate_report_views import *
